Move case/when expressions to expr crate
