+++ datafusion/expr/src/lib.rs
@@ -24,6 +24,7 @@ mod columnar_value;
 pub mod conditional_expressions;
 pub mod expr;
 pub mod expr_fn;
+pub mod expr_schema;
 pub mod field_util;
 pub mod function;
 mod literal;
@@ -43,12 +44,24 @@ pub use aggregate_function::AggregateFunction;
 pub use built_in_function::BuiltinScalarFunction;
 pub use columnar_value::{ColumnarValue, NullColumnarValue};
 pub use expr::Expr;
-pub use expr_fn::{col, sum};
+pub use expr_fn::{
+    abs, acos, and, approx_distinct, approx_percentile_cont, array, ascii, asin, atan,
+    avg, bit_length, btrim, case, ceil, character_length, chr, coalesce, col, concat,
+    concat_expr, concat_ws, concat_ws_expr, cos, count, count_distinct, date_part,
+    date_trunc, digest, exp, floor, in_list, initcap, left, length, ln, log10, log2,
+    lower, lpad, ltrim, max, md5, min, now, now_expr, nullif, octet_length, or, random,
+    regexp_match, regexp_replace, repeat, replace, reverse, right, round, rpad, rtrim,
+    sha224, sha256, sha384, sha512, signum, sin, split_part, sqrt, starts_with, strpos,
+    substr, sum, tan, to_hex, to_timestamp_micros, to_timestamp_millis,
+    to_timestamp_seconds, translate, trim, trunc, upper, when,
+};
+pub use expr_schema::ExprSchemable;
 pub use function::{
     AccumulatorFunctionImplementation, ReturnTypeFunction, ScalarFunctionImplementation,
     StateTypeFunction,
 };
 pub use literal::{lit, lit_timestamp_nano, Literal, TimestampLiteral};
+pub use logical_plan::{LogicalPlan, PlanVisitor};
 pub use nullif::SUPPORTED_NULLIF_TYPES;
 pub use operator::Operator;
 pub use signature::{Signature, TypeSignature, Volatility};
@@ -57,5 +70,3 @@ pub use udaf::AggregateUDF;
 pub use udf::ScalarUDF;
 pub use window_frame::{WindowFrame, WindowFrameBound, WindowFrameUnits};
 pub use window_function::{BuiltInWindowFunction, WindowFunction};
-
-pub use logical_plan::{LogicalPlan, PlanVisitor};
